requirements_2.txt
appdirs==1.4.4
astroid==2.4.2
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
DateTime==4.3
distlib==0.3.1
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Datepicker==0.14
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Security-Too==3.4.4
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
pdfkit==0.6.1
protobuf==3.13.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.6.0
PyMySQL==0.10.1
pytz==2020.4
selenium==3.141.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.20
toml==0.10.1
urllib3==1.26.2
virtualenv==20.0.35
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.3.3
zope.interface==5.2.0
